Move utils into src
