+++ commands/azureCommands/acr-logs.ts
@@ -6,8 +6,8 @@ import * as vscode from "vscode";
 import { IActionContext } from "vscode-azureextensionui";
 import { AzureImageTagNode, AzureRegistryNode } from '../../explorer/models/azureRegistryNodes';
 import { TaskNode } from "../../explorer/models/taskNode";
-import { getResourceGroupName, getSubscriptionFromRegistry } from '../../utils/Azure/acrTools';
-import { AzureUtilityManager } from '../../utils/azureUtilityManager';
+import { getResourceGroupName, getSubscriptionFromRegistry } from '../../src/utils/Azure/acrTools';
+import { AzureUtilityManager } from '../../src/utils/azureUtilityManager';
 import { quickPickACRRegistry } from '../utils/quick-pick-azure';
 import { accessLog } from "./acr-logs-utils/logFileManager";
 import { LogData } from "./acr-logs-utils/tableDataManager";
