+++ commands/azureCommands/create-registry.ts
@@ -8,9 +8,9 @@ import { SubscriptionModels } from 'azure-arm-resource';
 import { ResourceGroup } from "azure-arm-resource/lib/resource/models";
 import * as vscode from "vscode";
 import { ext } from '../../extensionVariables';
-import { isValidAzureName } from '../../utils/Azure/common';
-import { AzureUtilityManager } from '../../utils/azureUtilityManager';
-import { nonNullProp } from '../../utils/nonNull';
+import { isValidAzureName } from '../../src/utils/Azure/common';
+import { AzureUtilityManager } from '../../src/utils/azureUtilityManager';
+import { nonNullProp } from '../../src/utils/nonNull';
 import { quickPickLocation, quickPickResourceGroup, quickPickSKU, quickPickSubscription } from '../utils/quick-pick-azure';
 
 /* Creates a new Azure container registry based on user input/selection of features */
