+++ commands/azureCommands/run-task.ts
@@ -5,8 +5,8 @@ import { Subscription } from "azure-arm-resource/lib/subscription/models";
 import vscode = require('vscode');
 import { IActionContext, parseError } from "vscode-azureextensionui";
 import { TaskNode } from "../../explorer/models/taskNode";
-import * as acrTools from '../../utils/Azure/acrTools';
-import { AzureUtilityManager } from "../../utils/azureUtilityManager";
+import * as acrTools from '../../src/utils/Azure/acrTools';
+import { AzureUtilityManager } from "../../src/utils/azureUtilityManager";
 import { quickPickACRRegistry, quickPickSubscription, quickPickTask } from '../utils/quick-pick-azure';
 import { scheduleRunRequest } from '../utils/SourceArchiveUtility';
