+++ commands/utils/SourceArchiveUtility.ts
@@ -10,8 +10,8 @@ import * as tar from 'tar';
 import vscode = require('vscode');
 import { IActionContext, IAzureQuickPickItem } from 'vscode-azureextensionui';
 import { ext } from '../../extensionVariables';
-import { getBlobInfo, getResourceGroupName, streamLogs } from "../../utils/Azure/acrTools";
-import { AzureUtilityManager } from '../../utils/azureUtilityManager';
+import { getBlobInfo, getResourceGroupName, streamLogs } from "../../src/utils/Azure/acrTools";
+import { AzureUtilityManager } from '../../src/utils/azureUtilityManager';
 import { quickPickACRRegistry, quickPickSubscription } from './quick-pick-azure';
 import { Item, quickPickDockerFileItem, quickPickYamlFileItem } from './quick-pick-file';
 import { quickPickImageName } from './quick-pick-image';
