+++ configureWorkspace/configure_dotnetcore.ts
@@ -8,7 +8,7 @@ import * as path from 'path';
 import * as semver from 'semver';
 import { extractRegExGroups } from '../helpers/extractRegExGroups';
 import { isWindows, isWindows10RS3OrNewer, isWindows10RS4OrNewer, isWindows10RS5OrNewer } from '../helpers/osVersion';
-import { Platform, PlatformOS } from '../utils/platform';
+import { Platform, PlatformOS } from '../src/utils/platform';
 import { getExposeStatements, IPlatformGeneratorInfo, PackageInfo } from './configure';
 
 // This file handles both ASP.NET core and .NET Core Console
