+++ explorer/models/customRegistries.ts
@@ -7,7 +7,7 @@ import * as vscode from 'vscode';
 import { callWithTelemetryAndErrorHandling, IActionContext, parseError } from 'vscode-azureextensionui';
 import { keytarConstants } from '../../constants'
 import { ext } from '../../extensionVariables';
-import { nonNullValue } from '../../utils/nonNull';
+import { nonNullValue } from '../../src/utils/nonNull';
 import { CustomRegistryNode } from './customRegistryNodes';
 
 interface CustomRegistryNonsensitive {
