+++ explorer/models/registryRootNode.ts
@@ -10,10 +10,10 @@ import * as vscode from 'vscode';
 import { callWithTelemetryAndErrorHandling, IActionContext, parseError } from 'vscode-azureextensionui';
 import { keytarConstants, MAX_CONCURRENT_REQUESTS, MAX_CONCURRENT_SUBSCRIPTON_REQUESTS } from '../../constants';
 import { ext } from '../../extensionVariables';
+import { AsyncPool } from '../../src/utils/asyncpool';
+import { AzureUtilityManager } from '../../src/utils/azureUtilityManager';
+import { getLoginServer } from '../../src/utils/nonNull';
 import { AzureAccount } from '../../typings/azure-account.api';
-import { AsyncPool } from '../../utils/asyncpool';
-import { AzureUtilityManager } from '../../utils/azureUtilityManager';
-import { getLoginServer } from '../../utils/nonNull';
 import * as dockerHub from '../utils/dockerHubUtils';
 import { AzureLoadingNode, AzureNotSignedInNode, AzureRegistryNode } from './azureRegistryNodes';
 import { getCustomRegistries } from './customRegistries';
