+++ explorer/models/rootNode.ts
@@ -8,9 +8,9 @@ import * as vscode from 'vscode';
 import { callWithTelemetryAndErrorHandling, IActionContext } from 'vscode-azureextensionui';
 import { docker, ListContainerDescOptions as GetContainerDescOptions } from '../../commands/utils/docker-endpoint';
 import { ext, ImageGrouping } from '../../extensionVariables';
+import { AzureUtilityManager } from '../../src/utils/azureUtilityManager';
+import { treeUtils } from '../../src/utils/treeUtils';
 import { AzureAccount } from '../../typings/azure-account.api';
-import { AzureUtilityManager } from '../../utils/azureUtilityManager';
-import { treeUtils } from '../../utils/treeUtils';
 import { showDockerConnectionError } from '../utils/dockerConnectionError';
 import { ContainerNode, ContainerNodeContextValue } from './containerNode';
 import { ErrorNode } from './errorNode';
