+++ explorer/models/taskNode.ts
@@ -8,10 +8,10 @@ import * as ContainerModels from 'azure-arm-containerregistry/lib/models';
 import { SubscriptionModels } from 'azure-arm-resource';
 import * as vscode from 'vscode';
 import { callWithTelemetryAndErrorHandling, IActionContext } from 'vscode-azureextensionui';
+import * as acrTools from '../../src/utils/Azure/acrTools';
+import { AzureUtilityManager } from '../../src/utils/azureUtilityManager';
+import { treeUtils } from '../../src/utils/treeUtils';
 import { AzureAccount } from '../../typings/azure-account.api';
-import * as acrTools from '../../utils/Azure/acrTools';
-import { AzureUtilityManager } from '../../utils/azureUtilityManager';
-import { treeUtils } from '../../utils/treeUtils';
 import { openExternal } from '../utils/openExternal';
 import { NodeBase } from './nodeBase';
