+++ explorer/utils/browseAzurePortal.ts
@@ -4,8 +4,8 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { IActionContext } from 'vscode-azureextensionui';
+import { getTenantId, nonNullValue } from '../../src/utils/nonNull';
 import { AzureSession } from '../../typings/azure-account.api';
-import { getTenantId, nonNullValue } from '../../utils/nonNull';
 import { AzureImageTagNode, AzureRegistryNode, AzureRepositoryNode } from '../models/azureRegistryNodes';
 import { openExternal } from './openExternal';
