+++ extensionVariables.ts
@@ -10,7 +10,7 @@ import { ExtensionContext, OutputChannel } from "vscode";
 import { IAzureUserInput, ITelemetryReporter } from "vscode-azureextensionui";
 import { ITerminalProvider } from "./commands/utils/TerminalProvider";
 import { DockerExplorerProvider } from './explorer/dockerExplorerProvider';
-import { IKeytar } from './utils/keytar';
+import { IKeytar } from './src/utils/keytar';
 
 type requestPromise = RequestAPI<RequestPromise, RequestPromiseOptions, RequiredUriUrl>;
