+++ src/utils/Azure/acrTools.ts
@@ -14,10 +14,10 @@ import { ServiceClientCredentials } from 'ms-rest';
 import { TokenResponse } from 'ms-rest-azure';
 import { isNull } from 'util';
 import { parseError } from 'vscode-azureextensionui';
-import { NULL_GUID } from "../../constants";
-import { Manifest } from '../../explorer/utils/dockerHubUtils';
-import { ext } from '../../extensionVariables';
-import { AzureSession } from "../../typings/azure-account.api";
+import { NULL_GUID } from "../../../constants";
+import { Manifest } from '../../../explorer/utils/dockerHubUtils';
+import { ext } from '../../../extensionVariables';
+import { AzureSession } from "../../../typings/azure-account.api";
 import { AzureUtilityManager } from '../azureUtilityManager';
 import { getId, getLoginServer } from '../nonNull';
 import { AzureImage } from "./models/image";
