+++ src/utils/Azure/common.ts
