+++ src/utils/Azure/models/image.ts
