+++ src/utils/platform.ts
